[installer] Do not pull blobserve implementation into installer
