+++ components/blobserve/pkg/config/config.go
@@ -7,17 +7,15 @@ package config
 import (
 	"encoding/json"
 	"os"
-
-	"github.com/gitpod-io/gitpod/blobserve/pkg/blobserve"
 )
 
 // Config configures this service
 type Config struct {
-	BlobServe          blobserve.Config `json:"blobserve"`
-	AuthCfg            string           `json:"dockerAuth"`
-	PProfAddr          string           `json:"pprofAddr"`
-	PrometheusAddr     string           `json:"prometheusAddr"`
-	ReadinessProbeAddr string           `json:"readinessProbeAddr"`
+	BlobServe          BlobServe `json:"blobserve"`
+	AuthCfg            string    `json:"dockerAuth"`
+	PProfAddr          string    `json:"pprofAddr"`
+	PrometheusAddr     string    `json:"prometheusAddr"`
+	ReadinessProbeAddr string    `json:"readinessProbeAddr"`
 }
 
 // getConfig loads and validates the configuration
